requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
bs4==0.0.2
certifi==2025.10.5
chromedriver-autoinstaller==0.6.4
charset-normalizer==3.4.3
click==8.3.0
distro==1.9.0
dnspython==2.8.0
Flask==3.1.2
camelot-py==1.0.0
flask-cors==6.0.1
fpdf==1.7.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter
MarkupSafe==3.0.3
numpy==2.3.3
openai==2.3.0
outcome==1.3.0.post0
pandas==2.3.3
pydantic==2.12.0
pydantic_core==2.41.1
pymongo==4.15.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
selenium==4.36.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
tqdm==4.67.1
trio==0.31.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.9.0
webdriver-manager==4.0.2
waitress==3.0.2
Werkzeug==3.1.3
wsproto==1.2.0
